feat(bot): add Venture thread-per-task (thread per /task, ping inside)
When DISCORD_THREAD_ENABLED=true, a successful /task opens a Discord thread named "<ticket-id> <short title>" and pings the participants inside the thread — the main channel keeps only the short confirmation, so it never gets cluttered and each task becomes a jump-able sub-space.

- threads.py: pure helpers (build_thread_name / resolve_thread_pings / build_allowed_mentions / build_thread_intro) + best-effort open_task_thread that never raises into the handler.
- Participants: creator + assignee (resolved via reverse DISCORD_MEMBER_MAP, no backend call) + configured standing watchers.
- AllowedMentions scoped to exact ids (everyone=False) — no mass-ping possible.
- Public threads attach to the announcement message; private threads standalone.
- 6 new opt-in settings; default off preserves prior behaviour.
- Incidental: pre-existing mypy-2.0 type-only fixes so the mypy src gate is green.

Tests +50 (test_threads 35, handlers +7, config +8). ruff/mypy clean, 410 passed, coverage 80.38%. Version 0.2.7 -> 0.3.0.

+++ src/discord_agent_secretary/_worker.py
@@ -16,7 +16,9 @@ def backoff_seconds(base: float, failures: int, cap: float = BACKOFF_CAP_DEFAULT
     """Exponential backoff from `base`, doubling per consecutive failure, capped."""
     if failures <= 0:
         return base
-    return min(base * (2 ** (failures - 1)), cap)
+    # float() pins the result type: int**int is Any to mypy (negative exponent
+    # would yield float), which would otherwise leak Any out of this helper.
+    return float(min(base * (2 ** (failures - 1)), cap))
 
 
 def log_cycle_failure(
